+++ inc/classes/GoogleTranslate.php
@@ -203,14 +203,17 @@ class GoogleTranslate
     /**
      * Returns an array of Translations
      * 
-     * @param string      $text -> word to translate
-     * @param string      $sl -> source language code (i.e. en,de,fr,...)
-     * @param string      $tl -> target language code (i.e. en,de,fr,...)
-     * all supported language codes can be found here: https://cloud.google.com/translate/v2/using_rest#language-params
-     * @param string|null $time_token (optional) -> array() from https://translate.google.com
-     * if $time_token is empty, array(408254,585515986) is used
-     * @param string $domain (optional) -> connect to Google Domain (i.e. 'com' for  https://translate.google.com)
-     * if $domain is empty, a random domain will be used (the default value can be altered by changing DEFAULT_DOMAIN)
+     * @param string     $string     Word to translate
+     * @param string     $from       Source language code (i.e. en,de,fr,...)
+     * @param string     $to         Target language code (i.e. en,de,fr,...)
+     * all supported language codes can be found here: 
+     * https://cloud.google.com/translate/docs/basic/discovering-supported-languages#getting_a_list_of_supported_languages
+     * 
+     * @param int[]|null $time_token (optional) array() from 
+     * https://translate.google.com. If empty, array(408254,585515986) is used
+     * @param string     $domain     (optional) Connect to Google Domain 
+     * (i.e. 'com' for  https://translate.google.com). If empty, 
+     * a random domain will be used (the default value can be altered by changing DEFAULT_DOMAIN)
      * Possible values:
      *  ('com.ar', 'at', 'com.au', 'be', 'com.br', 'ca', 'cat', 'ch', 'cl', 'cn', 'cz', 
      * 'de', 'dk', 'es', 'fi', 'fr', 'gr', 'com.hk', 'hr', 'hu', 'co.id', 'ie',